requirements.txt

streamlit
numpy
pandas
scipy
plotly
scikit-learn
networkx
sympy
markdown
python-docx
requests
biopython
MDAnalysis
py3dmol
textstat
